* V5-6-patches:
  Fix 64-bit compiler warnings

+++ agent/mibgroup/ucd-snmp/loadave.c
@@ -393,7 +393,7 @@ write_laConfig(int action,
             double val;
             char *endp;
 
-            sprintf(buf, "%.*s", var_val_len, (char *)var_val);
+            sprintf(buf, "%.*s", (int) var_val_len, (char *)var_val);
             val = strtod(buf, &endp);
 
             if (errno == ERANGE || *endp != '\0' || val < 0 || val > 65536.00) {
